test: take floating point outputs scale
An output scale can be floating point. Provide this possibility in our test
library. And while at it we cleanup unnecessary calls to the function in our
tests having outputs with scale 1 only.

+++ autotests/integration/lib/app.h
@@ -68,7 +68,7 @@ public:
 
     void set_outputs(int count,
                      QVector<QRect> const& geometries = QVector<QRect>(),
-                     QVector<int> const& scales = QVector<int>());
+                     QVector<double> const& scales = QVector<double>());
 
 private:
     void handle_server_addons_created();
